requirements.txt
alembic==1.13.1
anyio==4.2.0
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.0
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.3.0
MarkupSafe==2.1.3
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot==20.7
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
typing_extensions==4.9.0
urllib3==2.2.0
Werkzeug==3.0.1
